Add x86_64-linux-gnu emulator
I wanted a tiny scriptable meltdown proof way to run userspace programs
and visualize how program execution impacts memory. It helps to explain
how things like Actually Portable Executable works. It can show you how
the GCC generated code is going about manipulating matrices and more. I
didn't feel fully comfortable with Qemu and Bochs because I'm not smart
enough to understand them. I wanted something like gVisor but with much
stronger levels of assurances. I wanted a single binary that'll run, on
all major operating systems with an embedded GPL barrier ZIP filesystem
that is tiny enough to transpile to JavaScript and run in browsers too.

https://justine.storage.googleapis.com/emulator625.mp4

+++ libc/str/sha256.c
@@ -31,7 +31,7 @@ static void sha256_transform(uint32_t state[hasatleast 8],
   size_t i;
   uint32_t a, b, c, d, e, f, g, h, t1, t2, m[64];
   for (i = 0; i < 16; ++i, data += 4) {
-    m[i] = data[0] << 24 | data[1] << 16 | data[2] << 8 | data[3];
+    m[i] = (uint32_t)data[0] << 24 | data[1] << 16 | data[2] << 8 | data[3];
   }
   for (; i < 64; ++i) {
     m[i] = SIG1(m[i - 2]) + m[i - 7] + SIG0(m[i - 15]) + m[i - 16];
